Move Groupalarm code into Networking module
